+++ src/vector_ops.cpp
@@ -10,8 +10,9 @@ Vector::Vector(size_t size, float default_value) : data_(size, default_value), s
 Vector::Vector(const std::initializer_list<float>& list) : data_(list), size_(list.size()) {}
 
 Vector::Vector(const Matrix& matrix) : data_(matrix.data(), matrix.data() + matrix.size()), size_(matrix.size()) {
+    // MSVC: loop vars must be signed int
     #pragma omp parallel for
-    for (size_t i = 0; i < matrix.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(matrix.size()); ++i) {
         data_[i] = matrix.at(i / matrix.cols(), i % matrix.cols());
     }
 }
@@ -23,8 +24,9 @@ void Vector::resize(size_t new_size) {
 }
 
 void Vector::fill(float value) {
+    // MSVC: loop vars must be signed int
     #pragma omp parallel for
-    for (size_t i = 0; i < size_; ++i) {
+    for (int i = 0; i < static_cast<int>(size_); ++i) {
         data_[i] = value;
     }
 }
@@ -33,8 +35,9 @@ void Vector::randomize(float min_val, float max_val) {
     std::random_device rd;
     std::mt19937 gen(rd());
     std::uniform_real_distribution<float> dis(min_val, max_val);
+    // MSVC: loop vars must be signed int
     #pragma omp parallel for
-    for (size_t i = 0; i < size_; ++i) {
+    for (int i = 0; i < static_cast<int>(size_); ++i) {
         float val = dis(gen);  // Note: Each thread gets its own random value
         data_[i] = val;
     }
@@ -44,16 +47,18 @@ void Vector::initialize_random(float scale) {
     std::random_device rd;
     std::mt19937 gen(rd());
     std::normal_distribution<float> dis(0.0f, scale);
+    // MSVC: loop vars must be signed int
     #pragma omp parallel for
-    for (size_t i = 0; i < size_; ++i) {
+    for (int i = 0; i < static_cast<int>(size_); ++i) {
         float val = dis(gen);  // Note: Each thread gets its own random value
         data_[i] = val;
     }
 }
 
 void Vector::initialize_constant(float value) {
+    // MSVC: loop vars must be signed int
     #pragma omp parallel for
-    for (size_t i = 0; i < size_; ++i) {
+    for (int i = 0; i < static_cast<int>(size_); ++i) {
         data_[i] = value;
     }
 }
@@ -62,8 +67,9 @@ Vector& Vector::operator+=(const Vector& other) {
     if (size_ != other.size()) {
         throw std::invalid_argument("Vector dimensions must match for addition");
     }
+    // MSVC: loop vars must be signed int
     #pragma omp parallel for
-    for (size_t i = 0; i < size_; ++i) {
+    for (int i = 0; i < static_cast<int>(size_); ++i) {
         data_[i] += other.data_[i];
     }
     return *this;
@@ -94,8 +100,9 @@ Vector operator-(const Vector& a, const Vector& b) {
         throw std::invalid_argument("Vector dimensions must match for subtraction");
     }
     Vector result(a.size());
+    // MSVC: loop vars must be signed int
     #pragma omp parallel for
-    for (size_t i = 0; i < a.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(a.size()); ++i) {
         result[i] = a[i] - b[i];
     }
     return result;
@@ -103,8 +110,9 @@ Vector operator-(const Vector& a, const Vector& b) {
 
 Vector operator*(const Vector& v, float scalar) {
     Vector result(v.size());
+    // MSVC: loop vars must be signed int
     #pragma omp parallel for
-    for (size_t i = 0; i < v.size(); ++i) {
+    for (int i = 0; i < static_cast<int>(v.size()); ++i) {
         result[i] = v[i] * scalar;
     }
     return result;
